将isize替换为usize | 索引bug修复 | JsonParser转换Dict时末键值对必须带逗号bug修复

+++ include/util/Dict.hpp
@@ -41,7 +41,7 @@ public:
      * 初始化一个空字典，使用指定的桶大小（默认为 MIN_BUCKET_SIZE）。
      * @param bucket_size 桶的初始大小。
      */
-    Dict(isize bucket_size = MIN_BUCKET_SIZE) :
+    Dict(usize bucket_size = MIN_BUCKET_SIZE) :
             bucket_(bucket_size), keys_() {}
 
     /**
@@ -104,7 +104,7 @@ public:
      * @brief 获取字典中键值对的数量。
      * @return 返回键值对的数量。
      */
-    isize size() const {
+    usize size() const {
         return keys_.size();
     }
 
@@ -120,7 +120,7 @@ public:
      * @brief 获取桶的容量。
      * @return 返回桶的容量。
      */
-    isize capacity() const {
+    usize capacity() const {
         return bucket_.capacity();
     }
 
@@ -370,7 +370,7 @@ public:
         if (this == &other) return *this;
         if (this->size() > other.size()) return other | *this;
 
-        Self res{static_cast<isize>((this->size() + other.size()) / MAX_LOAD_FACTOR)};
+        Self res{static_cast<usize>((this->size() + other.size()) / MAX_LOAD_FACTOR)};
         for (auto&& kv : *this) {
             res.insert_impl(kv.key(), kv.value(), my_hash(kv.key()));
         }
@@ -565,7 +565,7 @@ public:
          * @param dict 指向字典的指针。
          * @param index 初始索引。
          */
-        DictIterator(const Dict* dict, isize index) :
+        DictIterator(const Dict* dict, usize index) :
                 dict_(dict), index_(index) {
             update_kv(0);
         }
@@ -720,7 +720,7 @@ public:
          * @brief 更新当前键值对。
          * @param add 需要调整的指数。
          */
-        void update_kv(isize add) {
+        void update_kv(i64 add) {
             index_ += add;
             if (index_ < 0 || index_ >= dict_->size()) {
                 kv_.set(nullptr, nullptr);
@@ -732,7 +732,7 @@ public:
 
     private:
         const Dict* dict_; // 指向字典的指针
-        isize index_;      // 当前迭代器所在索引
+        usize index_;      // 当前迭代器所在索引
         value_type kv_;    // 当前键值对
     };
 
@@ -788,7 +788,7 @@ private:
      * 调用此函数将桶的大小扩大一倍。桶的初始大小为 `MIN_BUCKET_SIZE`，新的桶大小为当前大小的两倍。
      */
     void expand() {
-        bucket_.expand(std::max<isize>(MIN_BUCKET_SIZE, capacity() << 1LL));
+        bucket_.expand(std::max<usize>(MIN_BUCKET_SIZE, capacity() << 1LL));
     }
 
     /**
@@ -849,7 +849,7 @@ private:
     DynArray<key_t> keys_; // 键的动态数组
 
     constexpr static f64 MAX_LOAD_FACTOR = 0.75; // 最大负载因子
-    constexpr static isize MIN_BUCKET_SIZE = 8;  // 最小桶大小
+    constexpr static usize MIN_BUCKET_SIZE = 8;  // 最小桶大小
 };
 
 } // namespace my::util
