+++ tests/test_disjoint_set.hpp
@@ -29,7 +29,7 @@ fn should_merge_and_find = []() {
     Assertions::assertTrue(ds.is_connected(1, 6));
     Assertions::assertFalse(ds.is_connected(1, 8));
 
-    for (i32 i = 1; i <= 10; ++i) {
+    for (usize i = 1; i <= 10; ++i) {
         io::print(ds.rank(i));
     }
     io::println();
