+++ tests/test_vec.hpp
@@ -1,6 +1,7 @@
 #ifndef TEST_VEC_HPP
 #define TEST_VEC_HPP
 
+#include "Printer.hpp"
 #include "ricky_test.hpp"
 #include "Vec.hpp"
 
@@ -22,12 +23,12 @@ fn should_append = []() {
     util::Vec<CString> d;
 
     // When
-    for (i32 i = 0; i < n; ++i) {
+    for (usize i = 0; i < n; ++i) {
         d.append(cstr(i));
     }
 
     // Then
-    Assertions::assertEquals(isize(n), d.size());
+    Assertions::assertEquals(usize(n), d.size());
     Assertions::assertEquals(cstr(n - 1), d.at(d.size() - 1));
 
     i32 num = 0;
@@ -201,28 +202,28 @@ constexpr i32 N = 1e6;
 
 fn speed_of_vec_append_string = []() {
     util::Vec<std::string> d;
-    for (i32 i = 0; i < N; ++i) {
+    for (usize i = 0; i < N; ++i) {
         d.append("aaaaa");
     }
 };
 
 fn speed_of_vector_push_back_string = []() {
     std::vector<std::string> v;
-    for (i32 i = 0; i < N; ++i) {
+    for (usize i = 0; i < N; ++i) {
         v.push_back("aaaaa");
     }
 };
 
 fn speed_of_vec_append_i32 = []() {
     util::Vec<i32> d;
-    for (i32 i = 0; i < N; ++i) {
+    for (usize i = 0; i < N; ++i) {
         d.append(i);
     }
 };
 
 fn speed_of_vector_push_back_i32 = []() {
     std::vector<i32> v;
-    for (i32 i = 0; i < N; ++i) {
+    for (usize i = 0; i < N; ++i) {
         v.push_back(i);
     }
 };
